Factor out all redis code into lib
