+++ typescript/packages/toolshed/lib/behavior/strategies/scanForText.ts
@@ -1,5 +1,5 @@
-import { storage } from "../../../routes/storage/blobby/blobby.handlers.ts";
-import { getAllBlobs } from "../../../routes/storage/blobby/lib/redis.ts";
+import { getAllBlobs } from "@/lib/redis/redis.ts";
+import { storage } from "@/storage.ts";
 import { PrefixedLogger } from "../../prefixed-logger.ts";
 import { SearchResult } from "../search.ts";
