+++ typescript/packages/toolshed/storage.ts
@@ -1,0 +1,6 @@
+import { DiskStorage } from "@/lib/redis/storage.ts";
+
+const DATA_DIR = "./cache/blobby";
+
+export const storage = new DiskStorage(DATA_DIR);
+await storage.init();
